Move Gradle condition to Debug
