+++ graph/src/main/kotlin/net/twisterrob/gradle/graph/Debug.kt
@@ -1,0 +1,11 @@
+package net.twisterrob.gradle.graph
+
+object Debug {
+	/**
+	 * Setting this to the right value (fully qualified path of the project) enables clickable links from [WebView] logs.
+	 */
+	const val ResourcesFolder: String = """P:\projects\workspace\net.twisterrob.gradle\graph\src\main\resources\"""
+	const val JavaFx: Boolean = false
+	const val WebView: Boolean = true
+	const val Gradle: Boolean = true
+}
