--- graph/src/main/kotlin/net/twisterrob/gradle/graph/vis/d3/Debug.kt
@@ -1,10 +1,0 @@
-package net.twisterrob.gradle.graph.vis.d3
-
-object Debug {
-	/**
-	 * Setting this to the right value (fully qualified path of the project) enables clickable links from [WebView] logs.
-	 */
-	const val ResourcesFolder: String = """"""
-	const val JavaFx: Boolean = false
-	const val WebView: Boolean = false
-}
