+++ graph/src/main/kotlin/net/twisterrob/gradle/graph/vis/d3/GraphWindow.kt
@@ -7,6 +7,7 @@ import javafx.scene.layout.BorderPane
 import javafx.scene.paint.Color
 import javafx.scene.web.WebEngine
 import javafx.scene.web.WebView
+import net.twisterrob.gradle.graph.Debug
 import net.twisterrob.gradle.graph.tasks.TaskData
 import net.twisterrob.gradle.graph.tasks.TaskResult
 import net.twisterrob.gradle.graph.tasks.TaskType
